+++ scripts/covid dorset vaccinations.R
@@ -12,13 +12,14 @@ if (!require("pacman")) install.packages("pacman")
 pacman::p_load(
   ggplot2,
   tidyr,
-  reshape
+  reshape,
+  lubridate
 )
 
 # IMPORT DATASETS ----
 
-covid_vaccs_dor <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000059&metric=newPeopleVaccinatedFirstDoseByVaccinationDate&metric=newPeopleVaccinatedSecondDoseByVaccinationDate&format=csv"))
-covid_vaccs_bcp <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000058&metric=newPeopleVaccinatedFirstDoseByVaccinationDate&metric=newPeopleVaccinatedSecondDoseByVaccinationDate&format=csv"))
+covid_vaccs_dor <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000059&metric=newPeopleVaccinatedFirstDoseByVaccinationDate&metric=newPeopleVaccinatedSecondDoseByVaccinationDate&metric=newPeopleVaccinatedThirdInjectionByVaccinationDate&format=csv"))
+covid_vaccs_bcp <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000058&metric=newPeopleVaccinatedFirstDoseByVaccinationDate&metric=newPeopleVaccinatedSecondDoseByVaccinationDate&metric=newPeopleVaccinatedThirdInjectionByVaccinationDate&format=csv"))
 
 # MUNGE DATA ----
 
@@ -30,15 +31,19 @@ vaccs_list <- list(covid_vaccs_bcp,covid_vaccs_dor)
 vaccs_combined <- merge_recurse(vaccs_list)
 
 # keep only the columns we need, and rename the vacc event columns
-vaccs_combined <- subset(vaccs_combined, select = c("date", "areaName", "newPeopleVaccinatedFirstDoseByVaccinationDate", "newPeopleVaccinatedSecondDoseByVaccinationDate"))
+vaccs_combined <- subset(vaccs_combined, select = c("date", "areaName", "newPeopleVaccinatedFirstDoseByVaccinationDate", "newPeopleVaccinatedSecondDoseByVaccinationDate", "newPeopleVaccinatedThirdInjectionByVaccinationDate"))
 names(vaccs_combined)[names(vaccs_combined) == "newPeopleVaccinatedFirstDoseByVaccinationDate"] <- "First"
 names(vaccs_combined)[names(vaccs_combined) == "newPeopleVaccinatedSecondDoseByVaccinationDate"] <- "Second"
+names(vaccs_combined)[names(vaccs_combined) == "newPeopleVaccinatedThirdInjectionByVaccinationDate"] <- "Third or booster"
 
 # define the date format
 vaccs_combined$date = as.Date(vaccs_combined$date, "%Y-%m-%d")
 
+# restrict to last 9 months
+vaccs_combined <- subset(vaccs_combined, date > today() - months(6))
+
 # convert wide data into long
-vaccs_long <- gather(vaccs_combined, event, total, First, Second)
+vaccs_long <- gather(vaccs_combined, event, total, First:last_col())
 
 
 # PLOT DATA ----
@@ -46,14 +51,18 @@ vaccs_long <- gather(vaccs_combined, event, total, First, Second)
 # create plot and geom
 covid_vaccs_plot <- ggplot() +
   geom_bar(data = vaccs_long, aes(x = date, y = total, fill = event), stat="identity", position = "stack", width = 0.7) +
-  scale_fill_manual(name = "Vaccination", values = c("First" = "paleturquoise3", "Second" = "turquoise4"), labels = c("First", "Second")) +
+  scale_fill_manual(name = "Vaccination", values = c("First" = "paleturquoise3", "Second" = "turquoise4", "Third or booster" = "#00413d"), labels = c("First", "Second", "Third or booster")) +
   facet_grid( ~ areaName) +
-  theme_bw() +
-  scale_x_date(date_labels = "%B %Y", date_breaks = "2 months") +
+  scale_x_date(date_labels = "%b %y", date_breaks = "1 month") +
   ggtitle("Dorset covid vaccinations by local authority") +
-  labs(caption = paste("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted", Sys.time(), sep = " ")) +
+  labs(caption = paste("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted", Sys.time(), sep = " "), subtitle = "Daily numbers of new people that have received a COVID-19 vaccination") +
   xlab("Date") +
-  ylab("Vaccinations")
+  ylab("Vaccinations") +
+  theme_bw() +
+  theme(
+    plot.title = element_text(size = 20, family = "Helvetica", face = "bold"),
+    plot.subtitle = element_markdown(hjust = 0, vjust = 0, size = 11),
+    plot.caption = element_text(size = 10))
 
 covid_vaccs_plot