+++ scripts/covid dorset vaccs percentage.R
@@ -17,8 +17,8 @@ pacman::p_load(
 
 # IMPORT DATASETS ----
 
-vaccs_percentage_dor <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000059&metric=cumVaccinationFirstDoseUptakeByVaccinationDatePercentage&metric=cumVaccinationSecondDoseUptakeByVaccinationDatePercentage&format=csv"))
-vaccs_percentage_bcp <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000058&metric=cumVaccinationFirstDoseUptakeByVaccinationDatePercentage&metric=cumVaccinationSecondDoseUptakeByVaccinationDatePercentage&format=csv"))
+vaccs_percentage_dor <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000059&metric=cumVaccinationFirstDoseUptakeByVaccinationDatePercentage&metric=cumVaccinationSecondDoseUptakeByVaccinationDatePercentage&metric=cumVaccinationThirdInjectionUptakeByVaccinationDatePercentage&format=csv"))
+vaccs_percentage_bcp <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&areaCode=E06000058&metric=cumVaccinationFirstDoseUptakeByVaccinationDatePercentage&metric=cumVaccinationSecondDoseUptakeByVaccinationDatePercentage&metric=cumVaccinationThirdInjectionUptakeByVaccinationDatePercentage&format=csv"))
 
 # MUNGE DATA ----
 
@@ -30,30 +30,35 @@ vaccs_percentage_list <- list(vaccs_percentage_dor, vaccs_percentage_bcp)
 vaccs_percentage_combined <- merge_recurse(vaccs_percentage_list)
 
 # keep only the columns we need, and rename the vacc event columns
-vaccs_percentage_combined <- subset(vaccs_percentage_combined, select = c("date", "areaName", "cumVaccinationFirstDoseUptakeByVaccinationDatePercentage", "cumVaccinationSecondDoseUptakeByVaccinationDatePercentage"))
+vaccs_percentage_combined <- subset(vaccs_percentage_combined, select = c("date", "areaName", "cumVaccinationFirstDoseUptakeByVaccinationDatePercentage", "cumVaccinationSecondDoseUptakeByVaccinationDatePercentage", "cumVaccinationThirdInjectionUptakeByVaccinationDatePercentage"))
 names(vaccs_percentage_combined)[names(vaccs_percentage_combined) == "cumVaccinationFirstDoseUptakeByVaccinationDatePercentage"] <- "First"
 names(vaccs_percentage_combined)[names(vaccs_percentage_combined) == "cumVaccinationSecondDoseUptakeByVaccinationDatePercentage"] <- "Second"
+names(vaccs_percentage_combined)[names(vaccs_percentage_combined) == "cumVaccinationThirdInjectionUptakeByVaccinationDatePercentage"] <- "Third or booster"
 
 # define the date format
 vaccs_percentage_combined$date = as.Date(vaccs_percentage_combined$date, "%Y-%m-%d")
 
 # convert wide data into long
-vaccs_percentage_long <- gather(vaccs_percentage_combined, event, total, First:Second)
+vaccs_percentage_long <- gather(vaccs_percentage_combined, event, total, First:last_col())
 
 # PLOT DATA ----
 
 # create plot and geom
 covid_vaccs_percentage_plot <- ggplot() +
   geom_area(data = vaccs_percentage_long, aes(x = date, y = total, group = event, fill = event), position = "dodge") +
-  scale_fill_manual(name = "Vaccination", values = c("First" = "paleturquoise3", "Second" = "turquoise4"), labels = c("First", "Second")) +
+  scale_fill_manual(name = "Vaccination", values = c("First" = "paleturquoise3", "Second" = "turquoise4", "Third or booster" = "#00413d"), labels = c("First", "Second", "Third or booster")) +
   facet_grid( ~ areaName) +
-  theme_bw() +
-  scale_x_date(date_labels = "%B %Y", date_breaks = "2 months") +
+  scale_x_date(date_labels = "%b %y", date_breaks = "2 months") +
   ggtitle("Dorset covid vaccinations by local authority - population percentage over time") +
-  labs(caption = paste("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted", Sys.time(), sep = " ")) +
+  labs(caption = paste("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted", Sys.time(), sep = " "), subtitle = "Percentage of people aged 12 and over that have received a COVID-19 vaccination") +
   xlab("Date") +
   ylab("Vaccinations") +
-  scale_y_continuous(breaks = c(20, 40, 60, 80, 100), labels = scales::percent_format(scale = 1), limits = c(0,100))
+  scale_y_continuous(breaks = c(20, 40, 60, 80, 100), labels = scales::percent_format(scale = 1), limits = c(0,100)) +
+  theme_bw() +
+  theme(
+    plot.title = element_text(size = 20, family = "Helvetica", face = "bold"),
+    plot.subtitle = element_markdown(hjust = 0, vjust = 0, size = 11),
+    plot.caption = element_text(size = 10))
 
 covid_vaccs_percentage_plot